refactor: iterator 코드 분리
