+++ src/main/java/bowling/domain/NormalFrame.java
@@ -2,7 +2,6 @@ package bowling.domain;
 
 import java.text.MessageFormat;
 import java.util.Iterator;
-import java.util.Objects;
 
 public class NormalFrame extends AbstractFrame {
 
@@ -29,33 +28,7 @@ public class NormalFrame extends AbstractFrame {
 
     @Override
     public Iterator<Frame> iterator() {
-        final Frame me = this;
-        return new Iterator<Frame>() {
-
-            int frameNumber = 0;
-            Frame frame = me;
-
-            @Override
-            public boolean hasNext() {
-                if (frameNumber == 0) {
-                    return true;
-                }
-                return Objects.nonNull(frame.getNextFrame());
-            }
-
-            @Override
-            public Frame next() {
-                if (frameNumber == 0) {
-                    frameNumber++;
-                    return frame;
-                }
-                for (int i = 0; i < frameNumber; i++) {
-                    frame = frame.getNextFrame();
-                }
-                return frame;
-            }
-
-        };
+        return new FrameIterator(this);
     }
 
     @Override
